+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 
 ### Changed
 
+- Rename the `title` field to `question` in `type Poll` and `input PollInput`.
 - Rename `Mutation.setInvitability` to `Mutation.setPublicity`.
 - Add default values to `input GroupChatInput`.
 - Update `Query.readAccount` to take the ID of the user to read, and stop requiring an access token.